Conflicts:
	onemile/src/main/java/com/kh/onemile/vo/social/SocialDetailVO.java
	onemile/src/main/java/com/kh/onemile/vo/social/participate/ParticipateVO.java
	onemile/src/main/resources/mybatis/mapper/social/social-mapper.xml
	onemile/src/main/webapp/WEB-INF/views/social/detail.jsp

+++ onemile/src/main/java/com/kh/onemile/vo/social/SocialDetailVO.java
@@ -29,8 +29,8 @@ public class SocialDetailVO {
 	private List<ImageDTO> imageInfo; // 소셜의 이미지 목록 
 	//지도 테이블에서 보여줘야 할 내용들.
 	private int mapNo;
-	private double lat;
-	private double lng;
+	private float lat;
+	private float lng;
 	private String detailAddress;
 	
 }
